refactor(cli): move lane rail + milestones to tui_detail_rail.go (PRD #1009 M3)
